Change gcd from recursive to iterative
Note that, in the current version of rust, the new version and the
original recursive version performed *identically* in benchmarking
tests, suggesting that tail call optimization was indeed occurring.
So I suppose this is just "future-proofing".

Of course, I don't think there's much reason to think that the old
function would have ever stopped being optimized... so perhaps a
better word is "paranoia."

(I still hold that the recursive version looked far nicer)