Finer-grained trait bounds on `Factorization`
`Factorization` is one of the most central structs to the library,
so every trait bound it has ends up being required virtually
everywhere else in the library.  Some of these trait bounds have
been moved to the methods that explicitly require them.

Another alternative considered was to do the exact opposite: define
something like `Factorizable: Eq + Clone + Hash + Zero + One + Integer`,
and use it *everywhere*.  This would simplify a lot of the trait bounds
in the code, and should have little impact on the versatility of the
library seeing as I only really intend to support primitive integers
and BigInts in the first place.

...well drat, now I've convinced myself that I should've gone with the
alternative.  Oh well; save it for when I have more time to kill.