Document Factorization to expect prime factors
I've made a decision regarding composite factors in `Factorization`:
they are NOT okay, because `Factorization` implements mathematical
methods which only work as advertised when there are no composite
factors.

This decision was made with the acknowledgement that it is not
currently possible to statically guarantee that non-prime
`Factorization`s are never generated.