Write 'self' args in the conventional style
I've never really liked writing '&self' and '&mut self' as arguments,
because then I often end up accidentally writing the remaining
arguments in a similar style (e.g. `&x: i64` instead of `x: &i64`).

Because of this, at some point early on, I had decided that the `factor`
crate would flaunt that bit of convention and write `self` just like any
other argument. (the childish hope being that the rust team would soon
see the "error" of their ways and that `&self` would become deprecated)

In the time since, 1.0 has come out (hell, we're up to 1.6!), and
`&self` and `&mut self` are clearly here to stay. And probably with
good reason, considering that `self` IS special (it marks the difference
between a method and an associated function)... so I might as well start
following convention, huh?