Larger witness set --> Version 0.0.4
I need a larger witness set for what I'm doing right now, and
unfortunately, I no longer remember what the story is with respect
to nondeterministic testers in this crate.

While this is a terribly hacky way to meet my needs right now, the
only downside it has is regressions in performance, and as an upside,
the currently chosen upper bound is far less arbitrary than the
previous one.  This set of witnesses will give deterministic results
for all primitive integer types. (i.e. it works up to `::std::u64::MAX`)

In other words, the next time I do anything to this part of the
codebase, it'll be because I want to work with BigNums, and at that
point I'll be *forced* to start thinking about how add witness
selection and nondeterministic testing into the API.

Probably.