Add FactorExt
HashMaps for factorization are dumb.  That choice is nothing more
than a relic of decisions I made in ancient times when implementing
a similar thing in python.

This---together with some other silly decisions that preclude the
possibility of an efficient factorization sieve---has been slowly
pushing me away from this library back towards cross-crate copypasta.

Time to start fixing things.

Time for v0.2.