Factorization -> Factors
The intent was to replace Factorization with a pre-sorted Vec
structure.  At first I thought I could leverage existing code
in my `sparse` library (but I would've had to copy it due to
differing design constraints).  Then I thought I could find
an existing crate (I didn't).  So I settled for BTreeMap,
which has comparable performance to HashMap but will at least
allow me to remove all those Hash+Eq bounds.

I never added special cases for zero to the methods on `Factorization`,
and the more I think about it, the less I want to.  It'll be better
to just prohibit the factorization of zero, because it is ill-specified
to begin with.